Update Penjualan Jasa

+++ VPS_APP/process_unit/src/app/pesanan_penjualan_jasa/pesananPenjualanJasa.services.js
@@ -1,6 +1,9 @@
+import { formatDate } from "../../utils/jurnalUmumUtil.js"
 import { LOGGER, LOGGER_MONITOR, logType } from "../../utils/loggerUtil.js"
 import { generatePaginationResponse } from "../../utils/paginationUtil.js"
-import { createPesananPenjualanJasaRepo, deletePesananPenjualanJasaByUuidRepo, getAllPesananPenjualanJasaRepo, getPesananPenjualanJasaByUuidRepo, updatePesananPenjualanJasaByUuidRepo } from "./pesananPenjualanJasa.repository.js"
+import { getNeracaValidasiByTanggalService } from "../neraca/neraca.services.js"
+import { checkPerintahStokOpnameByNomorSuratPerintahAndBulanTransaksiService } from "../perintah_stok_opname/perintahStokOpname.services.js"
+import { createPesananPenjualanJasaRepo, deletePesananPenjualanJasaByUuidRepo, getAllPesananPenjualanJasaRepo, getPesananPenjualanJasaByUuidRepo, getTanggalTransaksiTerakhirByPesananPenjualanRepo, updatePesananPenjualanJasaByUuidRepo } from "./pesananPenjualanJasa.repository.js"
 
 export const getAllPesananPenjualanJasaService = async (query, req_identity) => {
     LOGGER(logType.INFO, "Start getAllPesananPenjualanJasaService", null, req_identity)
@@ -18,7 +21,7 @@ export const getAllPesananPenjualanJasaService = async (query, req_identity) => 
     LOGGER(logType.INFO, "Pagination", {
         pageNumber, size, search
     }, req_identity)
-    
+
     const pesananPenjualanJasas = await getAllPesananPenjualanJasaRepo(pageNumber, size, search, req_identity)
     return generatePaginationResponse(pesananPenjualanJasas.entry, pesananPenjualanJasas.count, pesananPenjualanJasas.pageNumber, pesananPenjualanJasas.size)
 }
@@ -40,13 +43,28 @@ export const createPesananPenjualanJasaService = async (pesananPenjualanJasaData
     LOGGER(logType.INFO, `Start createPesananPenjualanJasaService`, pesananPenjualanJasaData, req_identity)
     pesananPenjualanJasaData.enabled = 1
 
+    await getNeracaValidasiByTanggalService(null, pesananPenjualanJasaData.tanggal_pesanan_penjualan_jasa, req_identity)
+
+    await checkPerintahStokOpnameByNomorSuratPerintahAndBulanTransaksiService(null, pesananPenjualanJasaData.tanggal_pesanan_penjualan_jasa, null, null, req_identity)
+
     const pesananPenjualanJasa = await createPesananPenjualanJasaRepo(pesananPenjualanJasaData, req_identity)
     return pesananPenjualanJasa
 }
 
 export const deletePesananPenjualanJasaByUuidService = async (uuid, req_identity) => {
     LOGGER(logType.INFO, `Start deletePesananPenjualanJasaByUuidService [${uuid}]`, null, req_identity)
-    await getPesananPenjualanJasaByUuidService(uuid, req_identity)
+
+    const beforeData = await getPesananPenjualanJasaByUuidService(uuid, req_identity)
+
+    const allowToEdit = await getTanggalTransaksiTerakhirByPesananPenjualanService(uuid, beforeData.tanggal_pesanan_penjualan_jasa, req_identity);
+
+    if (!allowToEdit.allow) {
+        throw Error(JSON.stringify({
+            message: "Tidak Bisa Dihapus Karena Tanggal Terakhir Transaksi Adalah " + formatDate(allowToEdit.tanggal_minimum),
+            field: "error"
+        }))
+    }
+
     await deletePesananPenjualanJasaByUuidRepo(uuid, req_identity)
     return true
 }
@@ -54,6 +72,16 @@ export const deletePesananPenjualanJasaByUuidService = async (uuid, req_identity
 export const updatePesananPenjualanJasaByUuidService = async (uuid, pesananPenjualanJasaData, req_identity, req_original_url, req_method) => {
     LOGGER(logType.INFO, `Start updatePesananPenjualanJasaByUuidService [${uuid}]`, pesananPenjualanJasaData, req_identity)
     const beforeData = await getPesananPenjualanJasaByUuidService(uuid, req_identity)
+
+    const allowToEdit = await getTanggalTransaksiTerakhirByPesananPenjualanService(uuid, beforeData.tanggal_pesanan_penjualan_jasa, req_identity);
+
+    if (!allowToEdit.allow) {
+        throw Error(JSON.stringify({
+            message: "Tidak Bisa Diedit Karena Tanggal Terakhir Transaksi Adalah " + formatDate(allowToEdit.tanggal_minimum),
+            field: "error"
+        }))
+    }
+
     const pesananPenjualanJasa = await updatePesananPenjualanJasaByUuidRepo(uuid, pesananPenjualanJasaData, req_identity)
 
     LOGGER_MONITOR(req_original_url, req_method, {
@@ -62,4 +90,28 @@ export const updatePesananPenjualanJasaByUuidService = async (uuid, pesananPenju
     }, req_identity)
 
     return pesananPenjualanJasa
+}
+
+
+export const getTanggalTransaksiTerakhirByPesananPenjualanService = async (pesanan_penjualan, tanggal, req_identity) => {
+    LOGGER(logType.INFO, `Start getTanggalTransaksiTerakhirByPesananPenjualanService`, { pesanan_penjualan, tanggal }, req_identity)
+
+    const tanggalMinimum = await getTanggalTransaksiTerakhirByPesananPenjualanRepo(pesanan_penjualan, req_identity)
+
+    if (tanggalMinimum.length > 0) {
+        if (tanggalMinimum[0].tanggal_terakhir_transaksi == tanggal) {
+            return {
+                allow: true,
+                tanggal_minimum: tanggalMinimum[0].tanggal_terakhir_transaksi
+            }
+        } else {
+            return {
+                allow: false,
+                tanggal_minimum: tanggalMinimum[0].tanggal_terakhir_transaksi
+            }
+        }
+    }
+    return {
+        allow: false
+    };
 }
